starting work on advquery
